project/backend/requirements.txt
autopep8==1.5.7
click==7.1.2
fastapi==0.63.0
h11==0.12.0
mysql-connector-python==8.0.24
protobuf==3.15.8
pycodestyle==2.7.0
pydantic==1.8.1
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
